ML_visualizer/requirements.txt
numpy
matplotlib
PyOpenGL
glfw